fix jittering game camera

+++ src/com/sgrostad/gfx/GameCamera.java
@@ -38,8 +38,8 @@ public class GameCamera {
     }
 
     public void centerOnEntity(Entity e){
-        xOffset = e.getX() - handler.getWidth()/2 + e.getWidth()/2;
-        yOffset = e.getY() - handler.getHeight()/2 + e.getHeight()/2;
+        xOffset = (int)e.getX() - handler.getWidth()/2 + e.getWidth()/2;
+        yOffset = (int)e.getY() - handler.getHeight()/2 + e.getHeight()/2;
         avoidWhiteSpace();
     }
